Add OneOfHandler
No one uses this right now. We have to decide if the MVariable base-class is going to continue holding variables or not. If they are, then this class is probably not needed (it still could be used for separation of logic).

The trade-off is needing to have oneof logic throughout all generic code or oneof logic throughout each MVariable's code. Both seem not great.