Make expressions use int128 internally.
This is largely a speed boost and a corner-case catcher. We now do all arithmetic in 128 bit integers and validate when any operation is completed.

We also removed most absl::Status from evaluation, opting to return exceptions instead. When there's a failure, it's unlikely that any part of the stack wants to catch these errors.