Revamp Exactly and OneOf constructors
This should be more resilient for the short-term. These will be used largely in code, so we want a good user experience. This allows for OneOf({"abc", "def"}), which casts it to an std::string, so it is safe to store without them having to do type conversions or say `OneOf<std::string>({ ... })`.

We should probably create a typedef that resolves Moriarty types (e.g., vector<tuple<int, string>> -> vector<tuple<int64_t, string>>). But this needs more thought before we do it to see where it fits in (if anywhere).