Get 90% of MInteger into the new world
The only remaining item is the SizeProperty, which should be easy, but a lower priority for right now.

At this point, MInteger passes all constraints it knows about to MVariable<> so IsSatisfiedWith, MergeFrom, ToString can all be handled automatically.

* GetDependencies, ValueToString can also be automated soon as well

I'm not a big fan of the fact that we have a `RangeConstraint` wrapper, but I'm not sure if there's any other options. For right now, I'm just going to move on and we can make it prettier later!