Revamp Expressions
Two big fixes in one here. (Note: this commit is a breaking change on its own. The next commit fixes Expression's integration into the rest of the repo).

(1) No more absl::Status. If the expression doesn't parse, there is nothing to do but crash the program. So we're using exceptions instead.

(2) Our benchmarks show that std::visit was a non-negligible part of our runtime. ~18% of runtime in pathological cases. O_O. So we're using virtual classes now instead. Initial runs indicate it is faster, but we'll need true benchmarks to verify.