Remove MExampleGraph
We will likely bring this back in the future, but it's not worth the effort to maintain it right now. It is still in the old format, so when we bring it back, we will have to update it then. The documentation needs to be updated to walk through how to build constraints, etc.