Remove template argument of MVariable
This second argument can be inferred via type traits rather that explicitly stated. This leaves it in a pure CRTP state. In C++23, we might even get rid of that. This should hopefully help some compile error explosions. It may make other compile errors worse, but we'll see which is worse!